Refactored biome code
Switched to wrapping functions rather than accessing internal data.
The internal data seems to move around a lot making it difficult to update.
The rewrite stores a map from object pointer to id on creation and when the name is registered it writes to the file.
Added code to validate all the functions exist.